+++ src/app/App.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect} from 'react';
 import {useSelector} from 'react-redux';
-import PaperPicker from 'src/features/paperPicker/PaperPicker';
-import {getActivePaperId} from 'src/features/paperPicker/paperPickerSlice';
+import PaperPicker from 'src/features/paper/PaperPicker';
+import {getActivePaperId} from 'src/features/paper/paperSlice';
 import Reader from 'src/features/reader/Reader';
 
 /*
