--- src/common/util/types.ts
@@ -1,5 +1,0 @@
-/*
- * Utility types 
- */ 
-
-export type Optional<T> = T | undefined
