+++ src/features/reader/panels/main/Paragraph.tsx
@@ -1,20 +1,21 @@
 import {FunctionComponent} from "react"
 import {useSelector} from "react-redux"
-import {Paragraph as TParagraph, paragraphById} from "src/features/paperPicker/paperPickerSlice"
+import {Paragraph as TParagraph, ParagraphRef} from "src/common/types"
+import { paragraphById } from "src/features/paper/paperSlice"
 import ParagraphItem from "./ParagraphItem"
 
 interface ParagraphProps {
-	id: string
+	paragraphRef: ParagraphRef
 }
 
 /**
  * A paragraph of text in a paper.
  *
  * This is comprised of abstract "paragraph items", to allow for citation components to be embedded.
- * @param id 			The id of the paragraph to be rendered
+ * @param paragraphRef 			The paragraph to be rendered
  */
-const Paragraph: FunctionComponent<ParagraphProps> = ({ id }) => {
-	const paragraph = useSelector(paragraphById(id)) as TParagraph
+const Paragraph: FunctionComponent<ParagraphProps> = ({ paragraphRef }) => {
+	const paragraph = useSelector(paragraphById(paragraphRef.id)) as TParagraph
 
 	return (
 		<p>
