+++ src/features/reader/panels/main/ParagraphItem.tsx
@@ -1,8 +1,9 @@
 import {FunctionComponent} from "react"
-import {isCitationRef, isRefChain, isText} from "src/common/util/guards"
-import {ParagraphItem as TParagraphItem} from "src/features/paperPicker/paperPickerSlice"
+import {isCitationRef, isFigureRef, isRefChain, isText} from "src/common/util/guards"
+import {ParagraphItem as TParagraphItem} from "src/common/types"
 import Citation from "./Citation"
 import CitationChain from "./CitationChain"
+import Figure from "./Figure"
 
 interface ParagraphItemProps {
 	item: TParagraphItem
@@ -21,11 +22,12 @@ const ParagraphItem: FunctionComponent<ParagraphItemProps> = ({ item }) => {
 			</>
 		) 
 	} else if (isCitationRef(item)) {
-		return <Citation resourceRef={item} />
-
-		} else if (isRefChain(item)) {
-			return <CitationChain refChain={item} />
-		} else {
+		return <Citation citationRef={item} />
+	} else if (isRefChain(item)) {
+		return <CitationChain refChain={item} />
+	} else if (isFigureRef(item)) {
+		return <Figure figureRef={item} />
+	} else {
 		console.error(`Invalid section item: ${item}`)
 		return null
 	}
