+++ src/features/reader/panels/main/Section.tsx
@@ -1,7 +1,7 @@
 import {FunctionComponent} from "react";
 import FlexBox from "src/common/components/FlexBox";
-import {Section as TSection} from "src/features/paperPicker/paperPickerSlice";
-import SectionItem from "./SectionItem";
+import {Section as TSection} from "src/common/types";
+import Paragraph from "./Paragraph";
 
 interface SectionProps {
 	section: TSection
@@ -16,8 +16,8 @@ const Section: FunctionComponent<SectionProps> = ({ section }) => (
 	<FlexBox as="section" column>
 		<h2>{section.title}</h2>
 		{
-			section.content.map((item, i) => (
-				<SectionItem key={i} item={item} />
+			section.paragraphs.map((ref, i) => (
+				<Paragraph key={i} paragraphRef={ref} />
 			))
 		}
 	</FlexBox>
